+++ CHANGELOG.md
@@ -1,10 +1,16 @@
 ## 📒 ChangeLog
+### v1.3
+`2019-12-30`
+- ❌ Remove `active-color` is `Array`
+- 💄 Optimize the `color` for `Array`, supports up to 6 colors
+- 🎉 Added `color-flow` to set whether the color of the progress bar is flow
+
 ### v1.2
 `2019-12-28`
 - 💣 Attributes adjustment : `type` changed to `status`
 - 🎉 Added `border` to set whether the progress bar is circular
 - 💖 Added `lump` type
-  - 💜 Added `cut-width` to set `lump` width
+  - 💙 Added `cut-width` to set `lump` width
   - 🧡 Added `cut-color` to set `lump` color
 
 ### v1.1